// ui-react/apps/console/src/client/types.gen.ts
export type NamespaceTenantId = string;
// ...
export type DeviceUid = string;

/**
 * Device's status
 */
export type DeviceStatus = 'accepted' | 'rejected' | 'pending' | 'removed' | 'unused';

/**
 * Device's identity
 */
export type DeviceIdentity = {
    /**
     * Device's MAC
     */
    mac: string;
};

/**
 * Device's info
 */
export type DeviceInfo = {
    /**
     * Device's OS name
     */
    id: string;
    /**
     * Device's OS pretty name
     */
    pretty_name: string;
    /**
     * Device's OS version
     */
    version: string;
    /**
     * Device's OS arch
     */
    arch: string;
    /**
     * Device's OS platform
     */
    platform: 'docker' | 'native';
};

/**
 * Namespace's name
 */
export type NamespaceName = string;

/**
 * Device's status update date
 */
export type DeviceUpdateAt = string;

/**
 * A tag represents a label or category that can be attached to devices,
 * firewall rules and public keys for organization and filtering purposes.
 *
 */
export type Tag = {
    /**
     * The display name of the tag
     */
    name: string;
    /**
     * The tenant ID that owns this tag
     */
    tenant_id: string;
    /**
     * The timestamp when the tag was created
     */
    created_at: string;
    /**
     * The timestamp when the tag was last updated
     */
    updated_at: string;
};

/**
 * Device's Tags list
 */
export type DeviceTags = Array<Tag>;

/**
 * Device's public URL status.
 */
export type DevicePublicUrl = boolean;
// ...
export type DeviceAcceptable = boolean;

export type Device = {
    uid: DeviceUid;
    /**
     * Device's name
     *
     *
     * By default, the name is the device's MAC address when it just added.
     *
     */
    name: string;
    identity?: DeviceIdentity;
    info?: DeviceInfo;
    /**
     * Device's public key.
     */
    public_key?: string;
    tenant_id: NamespaceTenantId;
    /**
     * Device's last seen date
     */
    last_seen: string;
    /**
     * Device's availability status
     */
    online: boolean;
    namespace?: NamespaceName;
    status: DeviceStatus;
    status_update_at?: DeviceUpdateAt;
    /**
     * Device's creation date
     */
    created_at: string;
    /**
     * Device's remote address
     */
    remote_addr?: string;
    /**
     * Device's geolocation position
     */
    position?: {
        /**
         * Device's latitude position
         */
        latitude?: number;
        /**
         * Device's longitude position
         */
        longitude?: number;
    };
    tags?: DeviceTags;
    public_url?: DevicePublicUrl;
    acceptable?: DeviceAcceptable;
};
// ...
export type GetDevicesData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Filter field receives a JSON object enconded as base64 string for limit a search.
         *
         * The JSON enconded must follow these interafaces:
         * ```typescript
         * interface ParamProperty {
         * name: string;
         * operator: "contains" | "eq" | "bool" | "gt" | "lt";
         * value: string;
         * }
         *
         * interface ParamOperator {
         * name: "and" | "or";
         * }
         *
         * interface Filter {
         * type: "property" | "operator";
         * param: ParamOperator | ParamProperty;
         * }
         *
         * interface FilterList {
         * Filters: Array<Filter>;
         * }
         *
         * ```
         *
         * ## Examples
         *
         * This is a example to filter and get only the resource what property "confirmed" is "true"
         * ```json
         * [
         * {
         * "type": "property",
         * "params": {
         * "name": "confirmed",
         * "operator": "bool",
         * "value": "true"
         * }
         * }
         * ]
         * ```
         *
         * This one, filter resource by the property "id" inside "info" structure when it is equal to "manjaro" and online property is set to "true"
         * ```json
         * [
         * {
         * "type": "property",
         * "params": {
         * "name": "info.id",
         * "operator": "eq",
         * "value": "manjaro"
         * }
         * },
         * {
         * "type": "property",
         * "params": {
         * "name": "online",
         * "operator": "bool",
         * "value": "true"
         * }
         * },
         * {
         * "type": "operator",
         * "params": {
         * "name": "and"
         * }
         * }
         * ]
         * ```
         *
         */
        filter?: string;
        /**
         * Page number
         */
        page?: number;
        /**
         * Items per page
         */
        per_page?: number;
        /**
         * Device's status
         */
        status?: DeviceStatus;
        /**
         * Device's property to sort of
         */
        sort_by?: string;
        /**
         * Device's list order
         */
        order_by?: 'asc' | 'desc';
    };
    url: '/api/devices';
};